Move test somewhere it will actually get run
